move qt test calls to specific modules
