+++ src/core/ffi.rs
@@ -23,7 +23,7 @@
 
 use crate::core::jwt;
 use crate::core::timer::TimerWheel;
-use crate::version;
+use crate::core::version;
 use libc;
 use std::collections::HashSet;
 use std::env;
